Tuned parameters to work with more realistic membrane potentials
While some of the parameters of the model were in realistic ranges, the
membrane potentials were not (-5V to 2V). After setting the potentials
to be in more realistic ranges (-170mV to -30mV), some of the other
parameters were also tuned. Some changes were made to make more of the
variables set in the config (v_rest, exc_v_reset, inh_v_reset) to avoid
having to hard code it in model.py.